#65 fix: 시험 유형 관련 API의 도메인을 exam-type으로 변경
원래 lecture 도메인이였다.
근데 examType이 lecture_id와 관련 없이 되었으므로, exam-type으로 도메인을 따로 뺐다.
